requirements.txt
python-dotenv
openai
pandas
tqdm